Init JSON API fetcher and request constructor
