+++ .env
@@ -1,1 +1,2 @@
 VUE_APP_TITLE=Plenum Journal
+VUE_APP_CONTENTA_API='http://localhost:8888/contenta/web/api'
