+++ .env.development
@@ -1,2 +1,3 @@
 VUE_APP_TITLE=Plenum Journal (dev)
 VUE_APP_DRUPAL_HOSTNAME=http://localhost:8888/plenum-drupal-dev/drupal-8.5.3
+VUE_APP_CONTENTA_API=http://localhost:8888/contenta/web/api
