+++ package.json
@@ -12,6 +12,8 @@
   "dependencies": {
     "axios": "^0.18.0",
     "hsluv": "0.0.3",
+    "jsonapi-parse": "^2.0.0",
+    "typescript": "^3.1.6",
     "vue": "^2.5.17",
     "vue-check-view": "^0.2.3",
     "vue-class-component": "^6.0.0",
