+++ public/index.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <link rel="shortcut icon" href="<%= webpackConfig.output.publicPath %>favicon.ico">
+    <!--link rel="shortcut icon" href="<%= webpackConfig.output.publicPath %>favicon.ico"-->
     <link href="https://fonts.googleapis.com/css?family=Open+Sans:600|Amiri:400,400i" rel="stylesheet">
   </head>
   <body>
